requirements.txt
django
django-rest-framework
coreapi
django-cors-headers
gunicorn
whitenoise